fix sync test case
