+++ sync/src/block_connector/test_write_dag_block_chain.rs
@@ -7,7 +7,7 @@ use anyhow::{bail, Ok};
 use starcoin_account_api::AccountInfo;
 use starcoin_chain::{BlockChain, ChainReader};
 use starcoin_chain_service::WriteableChainService;
-use starcoin_config::{ChainNetwork, NodeConfig};
+use starcoin_config::{genesis_config::G_TEST_DAG_FORK_HEIGHT, ChainNetwork, NodeConfig};
 use starcoin_consensus::Consensus;
 use starcoin_crypto::HashValue;
 use starcoin_txpool_mock_service::MockTxPoolService;
@@ -128,7 +128,7 @@ pub fn new_dag_block(
 async fn test_dag_block_chain_apply() {
     let times = 12;
     let (mut writeable_block_chain_service, node_config, _) =
-        create_writeable_dag_block_chain(20).await;
+        create_writeable_dag_block_chain(G_TEST_DAG_FORK_HEIGHT).await;
     let net = node_config.net();
     let last_header_id = gen_dag_blocks(times, &mut writeable_block_chain_service, net);
     assert_eq!(
@@ -192,7 +192,7 @@ fn gen_fork_dag_block_chain(
 async fn test_block_dag_chain_switch_main() -> anyhow::Result<()> {
     let times = 12;
     let (mut writeable_block_chain_service, node_config, _) =
-        create_writeable_dag_block_chain(20).await;
+        create_writeable_dag_block_chain(G_TEST_DAG_FORK_HEIGHT).await;
     let net = node_config.net();
     let mut last_block = gen_dag_blocks(times, &mut writeable_block_chain_service, net)?;
     assert_eq!(
@@ -226,7 +226,7 @@ async fn test_block_dag_chain_switch_main() -> anyhow::Result<()> {
 async fn test_block_chain_reset() -> anyhow::Result<()> {
     let times = 10;
     let (mut writeable_block_chain_service, node_config, _) =
-        create_writeable_dag_block_chain(20).await;
+        create_writeable_dag_block_chain(G_TEST_DAG_FORK_HEIGHT).await;
     let net = node_config.net();
     let last_block = gen_dag_blocks(times, &mut writeable_block_chain_service, net)?;
     assert_eq!(
