+++ types/src/block/mod.rs
@@ -869,33 +869,33 @@ impl Block {
             );
         }
 
-        if !self.header.is_legacy() && !self.header.is_single() {
-            BlockMetadata::new_with_parents(
-                self.header.parent_hash(),
-                self.header.timestamp,
-                self.header.author,
-                self.header.author_auth_key,
-                uncles,
-                self.header.number,
-                self.header.chain_id,
-                parent_gas_used,
-                self.header
-                    .parents_hash
-                    .clone()
-                    .expect("Parents must exist"),
-            )
-        } else {
-            BlockMetadata::new(
-                self.header.parent_hash(),
-                self.header.timestamp,
-                self.header.author,
-                self.header.author_auth_key,
-                uncles,
-                self.header.number,
-                self.header.chain_id,
-                parent_gas_used,
-            )
-        }
+        // if !self.header.is_legacy() && !self.header.is_single() {
+        BlockMetadata::new_with_parents(
+            self.header.parent_hash(),
+            self.header.timestamp,
+            self.header.author,
+            self.header.author_auth_key,
+            uncles,
+            self.header.number,
+            self.header.chain_id,
+            parent_gas_used,
+            self.header
+                .parents_hash
+                .clone()
+                .expect("Parents must exist"),
+        )
+        // } else {
+        //     BlockMetadata::new(
+        //         self.header.parent_hash(),
+        //         self.header.timestamp,
+        //         self.header.author,
+        //         self.header.author_auth_key,
+        //         uncles,
+        //         self.header.number,
+        //         self.header.chain_id,
+        //         parent_gas_used,
+        //     )
+        // }
     }
 
     pub fn random() -> Self {
